Patch FreeImage to compile with mingw-w64

+++ src/FreeImage/Source/FreeImage/PluginTIFF.cpp
@@ -191,12 +191,8 @@ TIFFFdOpen(thandle_t handle, const char *name, const char *mode) {
 	    _tiffReadProc, _tiffWriteProc, _tiffSeekProc, _tiffCloseProc,
 	    _tiffSizeProc, _tiffMapProc, _tiffUnmapProc);
 
-	// Warning: tif_fd is declared as 'int' currently (see libTIFF), 
-    // may result in incorrect file pointers inside libTIFF on 
-    // 64bit machines (sizeof(int) != sizeof(long)). 
-    // Needs to be fixed within libTIFF.
     if (tif)
-		tif->tif_fd = (long)handle;
+		tif->tif_fd = (intptr_t)handle;
 
 	return tif;
 }
